FINISH HIM

+++ Cpp07/ex01/iter.hpp
@@ -4,11 +4,11 @@
 #include <iostream>
 
 template <typename T, typename F>
-void iter(T* array, int length, F func) 
+void iter(T* array, size_t length, F func) 
 {
     if (!array || !func)
         return;
-    for (int i = 0; i < length; i++) 
+    for (size_t i = 0; i < length; i++) 
         func(array[i]);
 }
